update kdr001- export excel N-O

+++ nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/scherec/monthlyattdcal/monthly/vacation/childcare/NursingCareLeaveMonthlyRemaining.java
@@ -33,5 +33,5 @@ public class NursingCareLeaveMonthlyRemaining {
     /**
      * 残時間:残時間　←子の看護休暇月別残数データ．本年使用数．使用数．時間
      */
-    private Double timeRemaining;
+    private Integer timeRemaining;
 }
